работает, но приходять не те цифры
