+++ src/components/BalanceNavigationSection/Balance/Balance.js
@@ -1,21 +1,40 @@
-import {React} from 'react';
+import React, {useState, useEffect} from 'react';
 // import { useSelector, useDispatch } from "react-redux";
 import Bubble from '../../Bubble/Bubble';
 import s from './Balance.module.css';
+import {getBalance} from "../../../redux/reports/reportsSelectors";
+import {fetchBalanceThunk} from "../../../redux/reports/reportsThunk";
+import {updateBalanceThunk} from "../../../redux/reports/reportsThunk";
+import { useDispatch, useSelector } from "react-redux";
 
-
-
-export default function Balance ({onInputChange, balance, onHandleSubmit}){
+export default function Balance (){
     
-    // const balance = 5000;
-    // const [balance, setBalance] = useState("");
-    // const dispatch = useDispatch();
-   
-   
+    const balance = useSelector(getBalance);
+    const dispatch = useDispatch();
+    
+    const [currentBalance, setCurrentBalance] = useState(Number(balance).toFixed(2));
     
     
-
-    // const initialZeroBalance = balance === 0;
+    
+    
+    const inputChange = (e) => { 
+       setCurrentBalance(e.target.value);
+    };
+    
+    
+    
+    const handleSubmit = (e) => {
+       e.preventDefault();
+       dispatch(updateBalanceThunk({ balance: currentBalance }),
+       [ balance]
+       )
+       
+    }
+    
+    useEffect(() => {
+       dispatch(fetchBalanceThunk())
+    setCurrentBalance(balance);
+           }, [balance, dispatch]);
 
    
 
@@ -25,21 +44,21 @@ export default function Balance ({onInputChange, balance, onHandleSubmit}){
 
     return (
         <div>
-            <form onSubmit={onHandleSubmit} className={s.balanceForm}> 
+            <form onSubmit={handleSubmit} className={s.balanceForm}> 
         
-        {balance === null ? (
+        {balance === 0 ? (
 
             <div className={s.inputWrapper}>
             <label htmlFor='balance' className={s.balanceLabel}>
                 
             <input 
-            type="text"
+            type="number"
             name="balance"
             placeholder={"00.00 UAH"}
             autoComplete="off"
-            onChange={onInputChange}
+            onChange={inputChange}
             className={s.balanceInput}
-            value={balance}
+            value={Number(currentBalance)}
             // value={balance}
             />
             
@@ -53,7 +72,7 @@ export default function Balance ({onInputChange, balance, onHandleSubmit}){
             <div className={s.activeBalanceContainer}>
                 
                 <div className={s.activeBalanceWrapper}>
-            <p className={`${s.balanceInput} ${s.balanceInputText}`}>{balance}.00 UAH</p>
+            <p className={`${s.balanceInput} ${s.balanceInputText}`}>{Number(currentBalance)}.00 UAH</p>
             
             </div>
             </div>
